% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microcancer-package.R
\docType{package}
\name{microcancer}
\alias{microcancer}
\alias{microcancer-package}
\title{microcancer: A package for micro-simulation for cancers}
\description{
microcancer: A package for micro-simulation for cancers
}
