% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{my_mvrnorm}
\alias{my_mvrnorm}
\title{This function uses mvrnorm from MASS}
\usage{
my_mvrnorm(m, s)
}
\arguments{
\item{m}{a numeric vector}

\item{s}{the covariance matrix}
}
\description{
This function uses mvrnorm from MASS
}
