% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{timesFive}
\alias{timesFive}
\title{This function multiply a numeric vector by 5}
\usage{
timesFive(x)
}
\arguments{
\item{x}{a numeric vector}
}
\description{
This function multiply a numeric vector by 5
}
